+++ src/parser/lexeme.cc
@@ -2,7 +2,8 @@
 
 /// Constructor for the Lexeme class
 Lexeme::Lexeme( const LexemeType type, const std::string charString,
-                const unsigned int startLineNo, const unsigned int startCharNo)
+                const unsigned long long int startLineNo,
+                const unsigned long long int startCharNo)
     : type(type),
       charString(charString),
       startLineNo(startLineNo),
@@ -20,11 +21,11 @@ LexemeType Lexeme::getType() {
 }
 
 /// Returns the line number of the first character of the lexeme
-unsigned int Lexeme::getStartLineNo() {
+unsigned long long int Lexeme::getStartLineNo() {
   return startLineNo;
 }
 
 /// Returns the character number of the first character of the lexeme
-unsigned int Lexeme::getStartCharNo() {
+unsigned long long int Lexeme::getStartCharNo() {
   return startCharNo;
 }
